backend/requirements.txt
Flask==2.0.1
Pillow==8.3.2
Flask-CORS==3.0.10